move all new entity collections into its own module and document
