+++ crates/re_viewer_context/src/lib.rs
@@ -17,6 +17,7 @@ mod space_view;
 mod store_context;
 mod tensor;
 mod time_control;
+mod typed_entity_collections;
 mod utils;
 mod viewer_context;
 
@@ -41,20 +42,22 @@ pub use selection_state::{
     Selection, SelectionHighlight,
 };
 pub use space_view::{
-    ApplicableEntitiesPerVisualizer, AutoSpawnHeuristic, DataResult, DynSpaceViewClass,
-    IdentifiedViewSystem, IndicatorMatchingEntities, IndicatorMatchingEntitiesPerVisualizer,
-    PerSystemDataResults, PerSystemEntities, PropertyOverrides, SpaceViewClass,
-    SpaceViewClassIdentifier, SpaceViewClassLayoutPriority, SpaceViewClassRegistry,
-    SpaceViewClassRegistryError, SpaceViewEntityHighlight, SpaceViewHighlights,
-    SpaceViewOutlineMasks, SpaceViewState, SpaceViewSystemExecutionError,
-    SpaceViewSystemRegistrator, SystemExecutionOutput, ViewContextCollection, ViewContextSystem,
-    ViewPartCollection, ViewPartSystem, ViewQuery, ViewSystemIdentifier, VisualizableEntities,
-    VisualizableEntitiesPerVisualizer, VisualizerAdditionalApplicabilityFilter,
-    VisualizerApplicableEntities,
+    AutoSpawnHeuristic, DataResult, DynSpaceViewClass, IdentifiedViewSystem, PerSystemDataResults,
+    PerSystemEntities, PropertyOverrides, SpaceViewClass, SpaceViewClassIdentifier,
+    SpaceViewClassLayoutPriority, SpaceViewClassRegistry, SpaceViewClassRegistryError,
+    SpaceViewEntityHighlight, SpaceViewHighlights, SpaceViewOutlineMasks, SpaceViewState,
+    SpaceViewSystemExecutionError, SpaceViewSystemRegistrator, SystemExecutionOutput,
+    ViewContextCollection, ViewContextSystem, ViewPartCollection, ViewPartSystem, ViewQuery,
+    ViewSystemIdentifier, VisualizerAdditionalApplicabilityFilter,
 };
 pub use store_context::StoreContext;
 pub use tensor::{TensorDecodeCache, TensorStats, TensorStatsCache};
 pub use time_control::{Looping, PlayState, TimeControl, TimeView};
+pub use typed_entity_collections::{
+    ApplicableEntitiesPerVisualizer, IndicatorMatchingEntities,
+    IndicatorMatchingEntitiesPerVisualizer, VisualizableEntities,
+    VisualizableEntitiesPerVisualizer, VisualizerApplicableEntities,
+};
 pub use utils::{auto_color, level_to_rich_text, DefaultColor};
 pub use viewer_context::{RecordingConfig, ViewerContext};
