+++ crates/re_viewer_context/src/space_view/mod.rs
@@ -29,15 +29,9 @@ pub use space_view_class_registry::{
 };
 pub use system_execution_output::SystemExecutionOutput;
 pub use view_context_system::{ViewContextCollection, ViewContextSystem};
-pub use view_part_system::{
-    ViewPartCollection, ViewPartSystem, VisualizableEntities, VisualizableEntitiesPerVisualizer,
-};
+pub use view_part_system::{ViewPartCollection, ViewPartSystem};
 pub use view_query::{DataResult, PerSystemDataResults, PropertyOverrides, ViewQuery};
 pub use visualizer_entity_subscriber::VisualizerAdditionalApplicabilityFilter;
-pub use visualizer_entity_subscriber::{
-    ApplicableEntitiesPerVisualizer, IndicatorMatchingEntities,
-    IndicatorMatchingEntitiesPerVisualizer, VisualizerApplicableEntities,
-};
 
 // ---------------------------------------------------------------------------
