+++ crates/re_viewer_context/src/space_view/space_view_class_registry.rs
@@ -5,11 +5,12 @@ use crate::{
     ApplicableEntitiesPerVisualizer, DynSpaceViewClass, IdentifiedViewSystem,
     IndicatorMatchingEntitiesPerVisualizer, SpaceViewClassIdentifier, ViewContextCollection,
     ViewContextSystem, ViewPartCollection, ViewPartSystem, ViewSystemIdentifier,
+    VisualizerApplicableEntities,
 };
 
 use super::{
     space_view_class_placeholder::SpaceViewClassPlaceholder,
-    visualizer_entity_subscriber::VisualizerEntitySubscriber, VisualizerApplicableEntities,
+    visualizer_entity_subscriber::VisualizerEntitySubscriber,
 };
 
 #[derive(Debug, thiserror::Error)]
